+++ src/theory/builtin/theory_builtin_type_rules.cpp
@@ -18,7 +18,6 @@
 #include "expr/attribute.h"
 #include "expr/skolem_manager.h"
 #include "expr/uninterpreted_constant.h"
-#include "util/cardinality.h"
 
 namespace cvc5 {
 namespace theory {
@@ -54,25 +53,6 @@ Node SortProperties::mkGroundTerm(TypeNode type)
       "groundTerm", type, "a ground term created for type " + type.toString());
   type.setAttribute(gta, k);
   return k;
-}
-
-Cardinality FunctionProperties::computeCardinality(TypeNode type)
-{
-  // Don't assert this; allow other theories to use this cardinality
-  // computation.
-  //
-  // Assert(type.getKind() == kind::FUNCTION_TYPE);
-
-  Cardinality argsCard(1);
-  // get the largest cardinality of function arguments/return type
-  for (size_t i = 0, i_end = type.getNumChildren() - 1; i < i_end; ++i)
-  {
-    argsCard *= type[i].getCardinality();
-  }
-
-  Cardinality valueCard = type[type.getNumChildren() - 1].getCardinality();
-
-  return valueCard ^ argsCard;
 }
 
 }  // namespace builtin
