+++ src/theory/datatypes/kinds
@@ -21,22 +21,22 @@ cardinality CONSTRUCTOR_TYPE \
 operator SELECTOR_TYPE 2 "selector"
 # can re-use function cardinality
 cardinality SELECTOR_TYPE \
-    "::cvc5::theory::builtin::FunctionProperties::computeCardinality(%TYPE%)" \
-    "theory/builtin/theory_builtin_type_rules.h"
+    "::cvc5::theory::uf::FunctionProperties::computeCardinality(%TYPE%)" \
+    "theory/uf/theory_uf_type_rules.h"
 
 # tester type has a constructor type
 operator TESTER_TYPE 1 "tester"
 # can re-use function cardinality
 cardinality TESTER_TYPE \
-    "::cvc5::theory::builtin::FunctionProperties::computeCardinality(%TYPE%)" \
-    "theory/builtin/theory_builtin_type_rules.h"
+    "::cvc5::theory::uf::FunctionProperties::computeCardinality(%TYPE%)" \
+    "theory/uf/theory_uf_type_rules.h"
 
 # tester type has a constructor type
 operator UPDATER_TYPE 2 "datatype update"
 # can re-use function cardinality
 cardinality UPDATER_TYPE \
-    "::cvc5::theory::builtin::FunctionProperties::computeCardinality(%TYPE%)" \
-    "theory/builtin/theory_builtin_type_rules.h"
+    "::cvc5::theory::uf::FunctionProperties::computeCardinality(%TYPE%)" \
+    "theory/uf/theory_uf_type_rules.h"
 
 parameterized APPLY_CONSTRUCTOR APPLY_TYPE_ASCRIPTION 0: "constructor application; first parameter is the constructor, remaining parameters (if any) are parameters to the constructor"
