+++ src/theory/uf/theory_uf_type_rules.cpp
@@ -19,6 +19,8 @@
 #include <sstream>
 
 #include "expr/cardinality_constraint.h"
+#include "theory/uf/function_const.h"
+#include "util/cardinality.h"
 #include "util/rational.h"
 
 namespace cvc5 {
@@ -158,6 +160,112 @@ TypeNode HoApplyTypeRule::computeType(NodeManager* nodeManager,
     }
     return nodeManager->mkFunctionType(children);
   }
+}
+
+TypeNode LambdaTypeRule::computeType(NodeManager* nodeManager,
+                                     TNode n,
+                                     bool check)
+{
+  if (n[0].getType(check) != nodeManager->boundVarListType())
+  {
+    std::stringstream ss;
+    ss << "expected a bound var list for LAMBDA expression, got `"
+       << n[0].getType().toString() << "'";
+    throw TypeCheckingExceptionPrivate(n, ss.str());
+  }
+  std::vector<TypeNode> argTypes;
+  for (TNode::iterator i = n[0].begin(); i != n[0].end(); ++i)
+  {
+    argTypes.push_back((*i).getType());
+  }
+  TypeNode rangeType = n[1].getType(check);
+  return nodeManager->mkFunctionType(argTypes, rangeType);
+}
+
+bool LambdaTypeRule::computeIsConst(NodeManager* nodeManager, TNode n)
+{
+  Assert(n.getKind() == kind::LAMBDA);
+  // get array representation of this function, if possible
+  Node na = FunctionConst::getArrayRepresentationForLambda(n);
+  if (!na.isNull())
+  {
+    Assert(na.getType().isArray());
+    Trace("lambda-const") << "Array representation for " << n << " is " << na
+                          << " " << na.getType() << std::endl;
+    // must have the standard bound variable list
+    Node bvl =
+        NodeManager::currentNM()->getBoundVarListForFunctionType(n.getType());
+    if (bvl == n[0])
+    {
+      // array must be constant
+      if (na.isConst())
+      {
+        Trace("lambda-const") << "*** Constant lambda : " << n;
+        Trace("lambda-const") << " since its array representation : " << na
+                              << " is constant." << std::endl;
+        return true;
+      }
+      else
+      {
+        Trace("lambda-const") << "Non-constant lambda : " << n
+                              << " since array is not constant." << std::endl;
+      }
+    }
+    else
+    {
+      Trace("lambda-const")
+          << "Non-constant lambda : " << n
+          << " since its varlist is not standard." << std::endl;
+      Trace("lambda-const") << "  standard : " << bvl << std::endl;
+      Trace("lambda-const") << "   current : " << n[0] << std::endl;
+    }
+  }
+  else
+  {
+    Trace("lambda-const") << "Non-constant lambda : " << n
+                          << " since it has no array representation."
+                          << std::endl;
+  }
+  return false;
+}
+
+Cardinality FunctionProperties::computeCardinality(TypeNode type)
+{
+  // Don't assert this; allow other theories to use this cardinality
+  // computation.
+  //
+  // Assert(type.getKind() == kind::FUNCTION_TYPE);
+
+  Cardinality argsCard(1);
+  // get the largest cardinality of function arguments/return type
+  for (size_t i = 0, i_end = type.getNumChildren() - 1; i < i_end; ++i)
+  {
+    argsCard *= type[i].getCardinality();
+  }
+
+  Cardinality valueCard = type[type.getNumChildren() - 1].getCardinality();
+
+  return valueCard ^ argsCard;
+}
+
+bool FunctionProperties::isWellFounded(TypeNode type)
+{
+  for (TypeNode::iterator i = type.begin(), i_end = type.end(); i != i_end; ++i)
+  {
+    if (!(*i).isWellFounded())
+    {
+      return false;
+    }
+  }
+  return true;
+}
+
+Node FunctionProperties::mkGroundTerm(TypeNode type)
+{
+  NodeManager* nm = NodeManager::currentNM();
+  Node bvl = nm->getBoundVarListForFunctionType(type);
+  Node ret = type.getRangeType().mkGroundTerm();
+  return nm->mkNode(kind::LAMBDA, bvl, ret);
 }
 
 }  // namespace uf
